requirements.txt
streamlit==1.36.0
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.6.1
imbalanced-learn==0.13.0
category-encoders==2.6.3
scipy==1.13.0
joblib==1.4.2
Pillow==10.3.0